pdf_mod/requirements.txt
requests
boto3==1.20.21
botocore==1.23.21
imageio==2.13.4
jmespath==0.10.0
networkx==2.6.3
numpy==1.21.5
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
PyPDF2==1.26.0
python-dateutil==2.8.2
PyWavelets==1.2.0
reportlab==3.6.3
s3transfer==0.5.0
six==1.16.0
tifffile==2021.11.2
urllib3==1.26.7
pyyaml
